+++ serviceProduct/src/main/java/com/amegdev/store/repository/IProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.amegdev.store.entity.Category;
 import com.amegdev.store.entity.Product;
 
-public interface IProductRepository extends JpaRepository<Product, Long> {
+public interface IProductRepository extends JpaRepository<Product, Integer> {
 
 	public List<Product> findByCategory(Category category);
